requirements.txt
dj-database-url==0.5.0
Django==2.1.3
djangorestframework==3.9.0
gunicorn==19.9.0
mysqlclient==1.3.12
python-decouple==3.1
pytz==2018.7
